refactor code layout
'lib' dirs under the 'handler' now named '_lib' according to convention of prefixing non-resource related directories.
